# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/recurringexpenses/{recurringExpenseId}:
    get:
      tags:
      - Recurring Expenses
      summary: Get recurring expense by ID
      description: Returns the specified recurring expense belonging to the authenticated
        user.
      operationId: getExpenseById
      parameters:
      - name: recurringExpenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Recurring expense retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
        "404":
          description: Recurring expense not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
    put:
      tags:
      - Recurring Expenses
      summary: Update a recurring expense
      description: Updates the specified recurring expense belonging to the authenticated
        user with the provided data.
      operationId: updateRecurringExpense
      parameters:
      - name: recurringExpenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringExpenseRequest"
        required: true
      responses:
        "200":
          description: Recurring expense updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
        "400":
          description: Invalid recurring expense data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
        "404":
          description: Recurring expense or category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
    delete:
      tags:
      - Recurring Expenses
      summary: Delete a recurring expense
      description: Deletes the specified recurring expense belonging to the authenticated
        user.
      operationId: deleteRecurringExpense
      parameters:
      - name: recurringExpenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Recurring expense deleted successfully
        "404":
          description: Recurring expense not found
  /api/expenses/{expenseId}:
    get:
      tags:
      - Expenses
      summary: Get expense by ID
      description: Returns the specified expense belonging to the authenticated user.
      operationId: getExpenseById_1
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Expense retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
        "404":
          description: Expense not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
    put:
      tags:
      - Expenses
      summary: Update an expense
      description: Updates the specified expense belonging to the authenticated user
        with the provided data.
      operationId: updateExpense
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseRequest"
        required: true
      responses:
        "200":
          description: Expense updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
        "400":
          description: Invalid expense data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
        "404":
          description: Expense or category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
    delete:
      tags:
      - Expenses
      summary: Delete an expense
      description: Deletes the specified expense belonging to the authenticated user.
      operationId: deleteExpense
      parameters:
      - name: expenseId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Expense deleted successfully
        "404":
          description: Expense not found
...
  /api/recurringexpenses:
    get:
      tags:
      - Recurring Expenses
      summary: Get all recurring expenses
      description: Returns all recurring expenses belonging to the authenticated user.
      operationId: getAllRecurringExpenses
      responses:
        "200":
          description: Recurring expense retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpenseResponse"
    post:
      tags:
      - Recurring Expenses
      summary: Create a recurring expense
      description: Creates a new recurring expense for the authenticated user.
      operationId: createRecurringExpense
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringExpenseRequest"
        required: true
      responses:
        "201":
          description: Recurring expense created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
        "400":
          description: Invalid recurring expense data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RecurringExpenseResponse"
...
  /api/recurringexpenses/search:
    get:
      tags:
      - Recurring Expenses
      summary: Search recurring expenses
      description: Returns the authenticated user's recurring expenses that match
        the specified search criteria.
      operationId: searchRecurringExpenses
      parameters:
      - name: title
        in: query
        required: false
        schema:
          type: string
      - name: frequency
        in: query
        required: false
        schema:
          type: string
          enum:
          - DAILY
          - WEEKLY
          - MONTHLY
          - YEARLY
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Recurring expenses retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpenseResponse"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpenseResponse"
  /api/recurringexpenses/filter:
    get:
      tags:
      - Recurring Expenses
      summary: Filter recurring expenses
      description: Returns the authenticated user's recurring expenses that match
        the specified filter criteria.
      operationId: filterRecurringExpenses
      parameters:
      - name: minAmount
        in: query
        required: true
        schema:
          type: number
      - name: maxAmount
        in: query
        required: true
        schema:
          type: number
      responses:
        "200":
          description: Recurring expense retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpenseResponse"
        "400":
          description: Invalid filtering criteria
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpenseResponse"
  /api/expenses/search:
    get:
      tags:
      - Expenses
      summary: Search expenses
      description: Returns the authenticated user's expenses that match the specified
        search criteria.
      operationId: searchExpenses
      parameters:
      - name: title
        in: query
        required: false
        schema:
          type: string
      - name: expenseDate
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: amount
        in: query
        required: false
        schema:
          type: number
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Expenses retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseResponse"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseResponse"
...
components:
  schemas:
    RecurringExpenseRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 0
        description:
          type: string
        amount:
          type: number
          exclusiveMinimum: 0
        categoryId:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        frequency:
          type: string
          enum:
          - DAILY
          - WEEKLY
          - MONTHLY
          - YEARLY
      required:
      - amount
      - categoryId
      - frequency
      - startDate
      - title
    RecurringExpenseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        amount:
          type: number
        categoryId:
          type: integer
          format: int32
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        frequency:
          type: string
          enum:
          - DAILY
          - WEEKLY
          - MONTHLY
          - YEARLY
    ExpenseRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 0
        description:
          type: string
        amount:
          type: number
          exclusiveMinimum: 0
        categoryId:
          type: integer
          format: int32
        expenseDate:
          type: string
          format: date
      required:
      - amount
      - categoryId
      - expenseDate
      - title
    ExpenseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        amount:
          type: number
        categoryId:
          type: integer
          format: int32
        expenseDate:
          type: string
          format: date
...